feat(sysmon): add option to use custom idle function (#5421)

+++ src/misc/lv_timer.c
@@ -258,7 +258,7 @@ void _lv_timer_core_deinit(void)
     _lv_ll_clear(timer_ll_p);
 }
 
-uint8_t lv_timer_get_idle(void)
+uint32_t lv_timer_get_idle(void)
 {
     return state.idle_last;
 }
